+++ adders/playwright/config/checks.ts
@@ -1,5 +1,5 @@
 import { defineAdderChecks } from '@svelte-add/core';
-import { options } from './options';
+import { options } from './options.js';
 
 export const checks = defineAdderChecks({
 	options,