+++ adders/playwright/index.ts
@@ -2,7 +2,7 @@
 
 import { defineAdder } from '@svelte-add/core';
 import { adder } from './config/adder.js';
-import { tests } from './config/tests.js';
 import { checks } from './config/checks.js';
+import { tests } from './config/tests.js';
 
 export default defineAdder(adder, checks, tests);