+++ eslint.config.js
@@ -33,6 +33,7 @@ export default tseslint.config(
 			'@typescript-eslint/no-unsafe-member-access': 'off',
 			'@typescript-eslint/no-unsafe-assignment': 'off',
 			'@typescript-eslint/no-empty-object-type': 'off',
+			'@typescript-eslint/no-unnecessary-type-parameters': 'off',
 		},
 	},
 	{