+++ packages/config/categories.ts
@@ -4,7 +4,7 @@ export type CategoryInfo = {
 	description: string;
 };
 
-export type CategoryKeys = 'codeQuality' | 'css' | 'db' | 'markdown' | 'tools';
+export type CategoryKeys = 'codeQuality' | 'css' | 'db' | 'testing' | 'additional';
 export type CategoryDetails = Record<CategoryKeys, CategoryInfo>;
 
 export type AdderCategories = Record<CategoryKeys, string[]>;
@@ -15,24 +15,24 @@ export const categories: CategoryDetails = {
 		name: 'Code Quality',
 		description: '',
 	},
+	testing: {
+		id: 'testing',
+		name: 'Testing',
+		description: '',
+	},
 	css: {
 		id: 'css',
 		name: 'CSS',
 		description: 'Can be used to style your components',
 	},
-	tools: {
-		id: 'tools',
-		name: 'Tools',
-		description: 'Different tools',
+	additional: {
+		id: 'additional',
+		name: 'Additional functionality',
+		description: '',
 	},
 	db: {
 		id: 'db',
 		name: 'Database',
-		description: '',
-	},
-	markdown: {
-		id: 'markdown',
-		name: 'Markdown',
 		description: '',
 	},
 };