+++ packages/testing-library/index.ts
@@ -1,7 +1,9 @@
+import path from 'node:path';
+import fs from 'node:fs';
 import { rm } from 'node:fs/promises';
 import { generateTestCases, runTestCases } from './utils/test-cases';
 import { getTemplatesDirectory } from './utils/workspace';
-import { downloadProjectTemplates } from './utils/create-project';
+import { downloadProjectTemplates, ProjectTypes } from './utils/create-project';
 import { remoteControl } from '@svelte-add/core/internal';
 import type { AdderWithoutExplicitArgs } from '@svelte-add/core/adder/config';
 
@@ -17,6 +19,21 @@ export async function testAdder(adder: AdderWithoutExplicitArgs, options: TestOp
 
 export async function testAdders(adders: AdderWithoutExplicitArgs[], options: TestOptions) {
 	await prepareTests(options);
+
+	const dirs: string[] = [];
+	for (const type of Object.values(ProjectTypes)) {
+		dirs.push(...adders.map((a) => `  - '${a.config.metadata.id}/${type}/*'`));
+	}
+
+	const pnpmWorkspace = `packages:\n${dirs.join('\n')}\n`;
+	fs.writeFileSync(path.join(options.outputDirectory, 'pnpm-workspace.yaml'), pnpmWorkspace, {
+		encoding: 'utf8',
+	});
+
+	const testRootPkgJson = JSON.stringify({ name: 'test-root', version: '0.0.0', type: 'module' });
+	fs.writeFileSync(path.join(options.outputDirectory, 'package.json'), testRootPkgJson, {
+		encoding: 'utf8',
+	});
 
 	remoteControl.enable();
 	await executeTests(adders, options);