+++ adders/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @svelte-add/adders
 
+## 1.6.0
+
+### Minor Changes
+
+- feat: remove bulma and bootstrap adders ([#533](https://github.com/svelte-add/svelte-add/pull/533))
+
+### Patch Changes
+
+- fix: removed `postcss` install from `tailwindcss` adder ([#532](https://github.com/svelte-add/svelte-add/pull/532))
+
+- fix: pinned `storybook` cli version ([#531](https://github.com/svelte-add/svelte-add/pull/531))
+
+## 1.5.2
+
+### Patch Changes
+
+- chore: use `defineConfig` helper for `playwright` ([#528](https://github.com/svelte-add/svelte-add/pull/528))
+
+## 1.5.1
+
+### Patch Changes
+
+- chore: upgrade to typescript-eslint v8 final ([#525](https://github.com/svelte-add/svelte-add/pull/525))
+
+## 1.5.0
+
+### Minor Changes
+
+- feat: playwright adder ([#518](https://github.com/svelte-add/svelte-add/pull/518))
+
+- feat: vitest adder ([#519](https://github.com/svelte-add/svelte-add/pull/519))
+
+## 1.4.0
+
+### Minor Changes
+
+- feat: generate `.env.example` file for drizzle adder ([#516](https://github.com/svelte-add/svelte-add/pull/516))
+
 ## 1.3.0
 
 ### Minor Changes