+++ packages/config/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @svelte-add/config
 
+## 1.5.2
+
+### Patch Changes
+
+- chore: fix repository urls ([#545](https://github.com/svelte-add/svelte-add/pull/545))
+
+## 1.5.1
+
+### Patch Changes
+
+- chore: publish package provenance info ([#541](https://github.com/svelte-add/svelte-add/pull/541))
+
+## 1.5.0
+
+### Minor Changes
+
+- feat: remove bulma and bootstrap adders ([#533](https://github.com/svelte-add/svelte-add/pull/533))
+
+## 1.4.0
+
+### Minor Changes
+
+- feat: playwright adder ([#518](https://github.com/svelte-add/svelte-add/pull/518))
+
+- feat: vitest adder ([#519](https://github.com/svelte-add/svelte-add/pull/519))
+
+### Patch Changes
+
+- chore: move `mdsvex` to a different category ([#521](https://github.com/svelte-add/svelte-add/pull/521))
+
 ## 1.3.0
 
 ### Minor Changes